Fix TS errors from Copilot review: extract cache to lib/cache.ts
- Move cache, loadCacheFromDisk, saveCacheToDisk to lib/cache.ts to
  avoid Next.js route export restrictions (reserved export names)
- Remove request.ip (not available on NextRequest in Next.js 15)
- Update cache.test.ts to import from lib/cache instead of route
